Added ability to receive a msg that is a multiple of MAX_MSG_SZ and more
* Changed BUFFER_SZ define in pub_socket_server.h to be more accurate.
* Renamed parameter variables for socket_server_send_data_all() and socket_server_send_data(). Removed const for data_sz.
* Added send_data() and send_data_all() private functions to prv_socket_server.h to be used internal to the library. Auto-inserts a command. Implemented these functions and updated socket_server_send_data_all() and socket_server_send_data() public functions to implement new functions. Pub functions use COMMAND_NONE. This was required to add in because otherwise user data send with these functions inserted no COMMAND and the message processing functions would have unexpected behavior as it parses for commands.
* Implemented above for socket client as well.
* Implemented ability for client/server to receive messages that are multiples of MAX_MSG_SZ in size. Note that it will only work if if the call to the send_data public functions for socket server/client are thread and instance safe (meaning that only one instance/thread can call these functions at a time or unexpected behavior will occur. This can't easily or reasonably be fixed, but this will not be a problem in the scope of this project due to the DBUS interface synchronously calling these functions.
* Changed socket_commons socket_receive_data function to return a flag instead of num bytes received and accept a ptr parameter for returning num bytes read. 
* Added SOCKET_RECEIVE_DATA_FLAGS enum as return values for socket_commons socket_receive_data().
* Change a few send/received function parameters/return value types to be larger/more appropiate.
* In socket_commons/pub_socket_commons.h, Added MSG_START__SEQUENCE and MSG_END__MORE as well as added command size and MSG size #defines. New message header types are used for receiving message that are multiples of MSG_MAX_SZ.
* In socket_commons changed remove_msg_header() to return SOCKET_RECEIVED_DATA_FLAGS enum values and accept a pointer to modify to represent the number of bytes a message is.
* socket_commons updated remove_msg_header()  function to determine if message received is expecting more messages of data, is the last message in a sequence, or is a standalone message.
* Added to socket_commons a new function socket_bytes_to_recv() that returns number of bytes to be read by socket.
* Fixed bug in socket_recv_data() where memcpy didn't account for the message header offsets cumulative to the loop number.

+++ Interfaces/1-Networking/socket_libs/client/pub_socket_client.h
@@ -82,7 +82,7 @@ void socket_client_execute();
  * 
  * @Returns number of bytes sent or RETURN_FAILED or 0 if there's an error.
  */
-int socket_client_send_data ( const char * data, const uint16_t data_sz );
+int socket_client_send_data ( const char * data, const uint data_sz );
 
 /**
  * Closes the socket connection to server and kills the
